docs: Add details about rpi password login/change
As a new user it's not immediately obvious what to put in when asked for the SSH Password. Adding a step describing what to enter and why eliminates the need for person to think about the fact that the script needs to login to the rpi and change the password when given the prompts.