Allow separately setting the overlay interface for MCP2515 kernel module
This is needed if MCP2515 is connected to spidev0.1 and spidev0.0 is used for some other SPI device. In this case mcp2515_overlay_can_device must be set to can1 or otherwise the MCP2515 kernel module uses spidev0.0. The network interface will still be can0 as there will only be one CAN interface in the system.