Added check for location of signalk-server
When provision script is run on ARMv6 architectures (Pi Zero and 1) it will install an unofficial node build and it will install signalk-server in /usr/local/bin instead of /usr/bin.
Added a check for /usr/bin/signalk-server and if not available, than use /usr/loca/bin in the template.